Add support for paired-end sequencing data (Fixes issue 9).
Paired-end reads are first judged by if the insert size is larger than
zero for any; If yes, then the reads are filtered if they are not properly
paired; The insert size is used instead of fragment size to calculate
coverage.

In addition, a debug tag is added as program argument.