Optimize the speed of RNA-seq coverage calculation.
In the original implementation, the returned data structure is coerced
to a data frame and then the data frames for all exons are rbind together.
This operation is very expensive in R. In this new implementation, the
coverage is calculated for each exon individually and put into a list.
The vectors corresponding to all exons are finally summed up into a single
vector. This enhancement has reduced running time by ~8%.