Further optimize RNA-seq coverage calculation.
In this new implementation, instead of scanning a bam file using multiple
exonic regions, a single DNA region is read. The coverage for the entire
DNA region is first calculated and then the exonic coverages are
concatenated. This is proved to be much faster than the original
implementation. Overall, I have managed to reduce the running time by ~40%.

The implementation also kills a previous unnoticed bug that counts a read
spanning multiple exons for more than once, resulting in elevated RPM.

The coverage calculation code is completely refactored to give much more elegant
solution than the previous version.