Merge covBam with covBamExons.
With the new implementation of RNA-seq coverage calculation, it has become
possible to calculate coverage for ChIP-seq using the same code. This
change reduces our burden of maintaining two functions that often need to
be modified together.

The merge introduces paired-end functionality into ChIP-seq too. However, I
noticed that the speed for ChIP-seq at TSS reduced by two folds. This is NOT
due to the extra computation on paired-end functionality. It remains a future
improvement to enhance the speed of the new covBamExons function.