Redefine boolean tag for paired-end data.
Previously, paired-end tag is set to True if any of the short read is
paired-end. Now, it is redefined as: True if all of the short reads
are paired-end. When all reads are either single-end or paired-end,
this makes no difference. But in the unlikely situation of mixed
paired-end and single-end reads, this is a safer resolution.