Fix zero outputs for some bam files.
This error was introduced when I started using "with" for lists.
However, when working with assignments to list members, the correct
way is to use "within". "with" will not change the list members. This
caused problems when the reads with mapping quality NA need to be updated
to 254: they were not updated and all of them were filtered out.