Remove the use of estiMapqStyle entirely
The estiMapqStyle function is not very well engineered. It has slowed
down program execution and caused trouble. So far, most aligners can
be properly recognized. In case there is a new aligner, it may still
be handled well if the aligner uses standard SAM format. Or else, a
user can always submit an issue report to us.